clean code & organize grpc api impl
