+++ Cargo.toml
@@ -193,6 +193,10 @@ wiremock = "0.5.18"
 url = "2.3.1"
 bs58 = "0.4.0"
 ring = "0.16.20"
+prost = "0.11.0"
+pbjson = "0.5"
+pbjson-types = "0.5"
+tonic = { version = "0.9.2", features = ["gzip"] }
 
 [dev-dependencies]
 cuckoofilter.workspace = true
