+++ crates/vrrb_grpc/Cargo.toml
@@ -15,12 +15,12 @@ proto_full = ["helloworld-v1"]
 # @@protoc_insertion_point(features)
 
 [dependencies]
-bytes = "1.1.0"
-prost = "0.11.0"
-pbjson = "0.5"
-pbjson-types = "0.5"
-serde = "1.0"
-tonic = { version = "0.9.2", features = ["gzip"] }
+bytes = { workspace = true }
+prost = { workspace = true }
+pbjson = { workspace = true }
+pbjson-types = { workspace = true }
+serde = { workspace = true }
+tonic = { workspace = true }
 anyhow = { workspace = true }
 mempool = { workspace = true }
 storage = { workspace = true }
