+++ crates/vrrb_grpc/buf.yaml
@@ -1,4 +1,5 @@
 version: v1
+name: buf.build/mattberryhill/vrrb
 breaking:
   use:
     - FILE
