+++ crates/vrrb_grpc/src/lib.rs
@@ -1,1 +1,2 @@
+pub mod handler;
 pub mod server;
